DROOLS-3226: [DMN Designer] Function Header Cell in BKM. Updates following peer review.
