+++ kie-wb-common-dmn/kie-wb-common-dmn-api/src/main/java/org/kie/workbench/common/dmn/api/definition/v1_1/BusinessKnowledgeModel.java
@@ -23,6 +23,7 @@ import org.jboss.errai.common.client.api.annotations.Portable;
 import org.jboss.errai.databinding.client.api.Bindable;
 import org.kie.soup.commons.util.Sets;
 import org.kie.workbench.common.dmn.api.definition.DMNViewDefinition;
+import org.kie.workbench.common.dmn.api.definition.HasExpression;
 import org.kie.workbench.common.dmn.api.definition.HasVariable;
 import org.kie.workbench.common.dmn.api.property.background.BackgroundSet;
 import org.kie.workbench.common.dmn.api.property.dimensions.RectangleDimensionsSet;
@@ -178,6 +179,30 @@ public class BusinessKnowledgeModel extends DRGElement implements HasVariable,
 
     public void setEncapsulatedLogic(final FunctionDefinition value) {
         this.encapsulatedLogic = value;
+    }
+
+    public HasExpression asHasExpression() {
+        return new HasExpression() {
+            @Override
+            public Expression getExpression() {
+                return BusinessKnowledgeModel.this.getEncapsulatedLogic();
+            }
+
+            @Override
+            public void setExpression(final Expression expression) {
+                throw new UnsupportedOperationException("It is not possible to set the EncapsulatedLogic of a BusinessKnowledgeModel.");
+            }
+
+            @Override
+            public DMNModelInstrumentedBase asDMNModelInstrumentedBase() {
+                return BusinessKnowledgeModel.this;
+            }
+
+            @Override
+            public boolean isClearSupported() {
+                return false;
+            }
+        };
     }
 
     // ------------------------------------------------------
